[refactor] Domain Layer model을 개선합니다.
